requirements.txt
asgiref==3.5.0
colorama==0.4.4
Django==3.2.11
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-api-key==2.1.0
djangorestframework-simplejwt==5.0.0
Pillow==9.0.0
PyJWT==2.3.0
pyotp==2.6.0
pytz==2021.3
pywin32==303
qrcode==7.3.1
sqlparse==0.4.2
typing_extensions==4.0.1
psycopg2==2.9.3